# internal/keeper/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Keeper API",
    "description": "API spec for Tacito Square Keeper component, managing agent configurations, communities, and support entities.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Primary API Version 1"
    }
  ],
...
  "paths": {
    "/llm-bindings": {
      "post": {
        "summary": "Create a new LLM provider binding",
        "operationId": "createLLMBinding",
        "tags": [
          "agent-config/llm-bindings"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLLMBindingRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created LLM provider binding",
            "headers": {
              "Location": {
                "description": "The URI of the newly created llm bindings",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request / Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "List all LLM provider bindings",
        "operationId": "listLLMBindings",
        "tags": [
          "agent-config/llm-bindings"
        ],
        "responses": {
          "200": {
            "description": "List of LLM provider bindings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LLMBinding"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateLLMBindingRequest": {
        "type": "object",
        "required": [
          "name",
          "provider",
          "api_base_url",
          "api_key_secret_ref",
          "default_model"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "openai-gpt4o"
          },
          "description": {
            "type": "string",
            "example": "Primary OpenAI GPT-4o endpoint"
          },
          "provider": {
            "type": "string",
            "enum": [
              "openai",
              "anthropic",
              "groq",
              "ollama",
              "custom"
            ],
            "example": "openai"
          },
          "api_base_url": {
            "type": "string",
            "format": "uri",
            "example": "https://api.openai.com/v1"
          },
          "api_key_secret_ref": {
            "type": "string",
            "example": "openai-api-key"
          },
          "default_model": {
            "type": "string",
            "example": "gpt-4o"
          },
          "default_temperature": {
            "type": "number",
            "format": "double",
            "minimum": 0.0,
            "maximum": 2.0,
            "default": 0.7,
            "example": 0.7
          },
          "default_max_tokens": {
            "type": "integer",
            "minimum": 1,
            "default": 2048,
            "example": 2048
          },
          "timeout_seconds": {
            "type": "integer",
            "minimum": 1,
            "default": 30,
            "example": 30
          }
        }
      },
...
      "LLMBinding": {
        "type": "object",
        "required": [
          "id",
          "name",
          "provider",
          "api_base_url",
          "api_key_secret_ref",
          "default_model",
          "default_temperature",
          "default_max_tokens",
          "timeout_seconds",
          "status",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "provider": {
            "type": "string",
            "enum": [
              "openai",
              "anthropic",
              "groq",
              "ollama",
              "custom"
            ]
          },
          "api_base_url": {
            "type": "string",
            "format": "uri"
          },
          "api_key_secret_ref": {
            "type": "string"
          },
          "default_model": {
            "type": "string"
          },
          "default_temperature": {
            "type": "number",
            "format": "double"
          },
          "default_max_tokens": {
            "type": "integer"
          },
          "timeout_seconds": {
            "type": "integer"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "suspended",
              "inactive"
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "example": "name is required"
          }
        }
      },
...
    }
  }
}